+++ packages/wordpress-kit/CHANGELOG.md
@@ -10,7 +10,8 @@
 
 ### Major Changes
 
-- 5dfbb69: [wordpress-kit] We move to use graphql-request instead of apollo client, updated the test for that, the main example and the exports
+- 5dfbb69: [wordpress-kit] We move to use graphql-request instead of apollo
+  client, updated the test for that, the main example and the exports
 
   ### Breaking